move event_catcher fixture to be reusable
