+++ pyproject.toml
@@ -130,7 +130,7 @@ exclude = [
     "env*",
     ".hatch/*",
 ]
-per-file-ignores = ["*/__init__.py: F401"]
+per-file-ignores = ["*/__init__.py: F401", "*/conftest.py: F401"]
 docstring-convention = "google"
 
 [tool.mypy]
