+++ tests/unit/conftest.py
@@ -1,0 +1,1 @@
+from tests.unit.utils import event_catcher
